Modified ctrlin/out64, added ld24 & CTRL_SYSTEM1
Removed the base register parameter for ctrlin64 and ctrlout64, mostly because this makes it easier to save the program's registers in an operating system (otherwise it'd have to thrash a register). The downside of this is we now can't as easily get a control register dynamically.
Added an optimised ld24 set of instructions (these just load and sign-extend a 24-bit immediate value into one of the lower 16 registers).
Added an additional "SYSTEM1" control register (this is another register like SYSTEM0 which can hold system data or temporary values).