UNTESTED: Added a real-time MMU
The power went out last night, so I added an MMU instead of researching other stuff.
This is only about 350 lines of code (and an extra 16 internal control registers), so it adds quite a bit to the weight on FPGA (and you might want to replace it with a dummy zero-slot MMU for smaller implementations) but in terms of logic size it should be manageable, assuming it works properly.
This MMU does NOT do any cacheing and it does NOT rely on external memory for any kind of page table (rather, it has 8 slots which can each map a 1KB-aligned page up to 64MB in size with options e.g. whether it's only allowed from system more or only for read or write). So given that it has less features than a regular MMU but should operate with more reliable timing, the proper description would probably be "real-time MMU".