UNTESTED: It even has register protection now
The idea behind being able to limit access to higher general-purpose registers is to aid in multitasking for secure workloads: Your high-security (e.g. cryptography) task can use the higher registers for it's secrets, with lower-priority tasks being limited to lower registers. This means the higher registers don't need to be swapped out to RAM in order to switch tasks, so your cryptography secrets can be kept extra-extra-extra safe.